backend/requirements.txt
flask==3.0.3
flask-smorest==0.44.0
flask-cors==4.0.0
pydantic==2.9.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
alembic==1.13.2
argon2-cffi==23.1.0
pyjwt[crypto]==2.9.0
redis==5.0.8
celery==5.4.0
boto3==1.35.25
minio==7.2.10
qdrant-client==1.11.1
requests==2.32.3
uvicorn==0.30.6
gunicorn==22.0.0
werkzeug==3.0.4
structlog==24.4.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
